VistasHQ/requirements.txt
asgiref==3.6.0
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7
django-import-export==4.1.1
django-jazzmin==3.0.0
django-crispy-forms==2.3
django-mathfilters==1.0.0
django-humanize==0.1.2
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-taggit==6.0.0
djangorestframework-simplejwt==2.1
django-crispy-forms
